Moving code from service to use-case
