+++ src/core/application/use-cases/issues/get-issue-by-id.use-case.ts
@@ -6,6 +6,8 @@ import { IUseCase } from '@/shared/domain/interfaces/use-case.interface';
 import { Inject, Injectable } from '@nestjs/common';
 import { BuildFilterUseCase } from './build-filter.use-case';
 import { IssuesEntity } from '@/core/domain/issues/entities/issues.entity';
+import { CODE_REVIEW_FEEDBACK_SERVICE_TOKEN, ICodeReviewFeedbackService } from '@/core/domain/codeReviewFeedback/contracts/codeReviewFeedback.service.contract';
+import { IIssue } from '@/core/domain/issues/interfaces/issues.interface';
 
 @Injectable()
 export class GetIssueByIdUseCase implements IUseCase {
@@ -13,10 +15,63 @@ export class GetIssueByIdUseCase implements IUseCase {
         @Inject(ISSUES_SERVICE_TOKEN)
         private readonly issuesService: IIssuesService,
 
+        @Inject(CODE_REVIEW_FEEDBACK_SERVICE_TOKEN)
+        private readonly codeReviewFeedbackService: ICodeReviewFeedbackService,
+
         private readonly buildFilterUseCase: BuildFilterUseCase,
     ) {}
 
-    async execute(id: string): Promise<IssuesEntity> {
-        return this.issuesService.findById(id);
+    async execute(id: string): Promise<IIssue | null> {
+        const issue = await this.issuesService.findById(id);
+
+        if (!issue) {
+            return null;
+        }
+
+        const codeReviewFeedback =
+            await this.codeReviewFeedbackService.getByOrganizationId(
+                issue.organizationId,
+            );
+
+        if (!codeReviewFeedback) {
+            return issue;
+        }
+
+        const issueFeedbacks = codeReviewFeedback.filter(
+            (feedback) =>
+                feedback?.suggestionId === issue.representativeSuggestion.id,
+        );
+
+        const prNumbers = await this.selectAllPrNumbers(issue);
+
+        const issueWithFeedback = {
+            ...issue.toObject(),
+            reactions: issueFeedbacks.map((feedback) => feedback.reactions),
+            prNumbers,
+        };
+
+        return issueWithFeedback;
+    }
+
+    private async selectAllPrNumbers(issue: IssuesEntity): Promise<string[]> {
+        const prNumbers = new Set<string>();
+
+        if (issue.representativeSuggestion?.prNumber) {
+            prNumbers.add(issue.representativeSuggestion.prNumber.toString());
+        }
+
+        if (issue.contributingSuggestions?.length) {
+            issue.contributingSuggestions.forEach((suggestion) => {
+                if (suggestion.prNumber) {
+                    prNumbers.add(suggestion.prNumber.toString());
+                }
+            });
+        }
+
+        const orderedPrNumbers = new Set(
+            Array.from(prNumbers).sort((a, b) => parseInt(a) - parseInt(b)),
+        );
+
+        return Array.from(orderedPrNumbers);
     }
 }
