+++ src/core/domain/issues/interfaces/issues.interface.ts
@@ -19,4 +19,9 @@ export interface IIssue {
     status: IssueStatus;
     createdAt: string;
     updatedAt: string;
+    prNumbers?: string[];
+    reactions?: {
+        thumbsUp: number;
+        thumbsDown: number;
+    }[];
 }
